Moved all the locations API services to locations; moved all the locations urls to locations_build_urls
