+++ bingmaps/apiservices/__init__.py
@@ -1,0 +1,5 @@
+from .locations import (
+    LocationByAddress,
+    LocationByPoint,
+    LocationByQuery
+)
