--- bingmaps/location/__init__.py
@@ -1,3 +1,0 @@
-from .location_by_address import LocationByAddress
-from .location_by_point import LocationByPoint
-from .location_by_query import LocationByQuery
