+++ bingmaps/urls/__init__.py
@@ -1,0 +1,11 @@
+from .locations_build_urls import (
+    LocationByAddressQueryString,
+    LocationByAddressSchema,
+    LocationByAddressUrl,
+    LocationByPointQueryString,
+    LocationByPointSchema,
+    LocationByPointUrl,
+    LocationByQuerySchema,
+    LocationByQueryString,
+    LocationByQueryUrl
+)
