--- bingmaps/urlschema/__init__.py
@@ -1,17 +1,0 @@
-from .location_by_address_schema import (
-    LocationByAddressSchema,
-    LocationByAddressQueryString,
-    LocationByAddressUrl
-)
-
-from .location_by_point_schema import (
-    LocationByPointQueryString,
-    LocationByPointSchema,
-    LocationByPointUrl
-)
-
-from .location_by_query_schema import (
-    LocationByQuerySchema,
-    LocationByQueryString,
-    LocationByQueryUrl
-)
